+++ tunnel/koupleless_mqtt_tunnel/consts.go
@@ -1,4 +1,4 @@
-package mqtt_tunnel
+package koupleless_mqtt_tunnel
 
 const (
 	CommandHealth       = "health"
