+++ vnode/base_node/pod_provider/provider_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/koupleless/arkctl/v1/service/ark"
 	"github.com/koupleless/virtual-kubelet/common/mqtt"
 	"github.com/koupleless/virtual-kubelet/common/testutil/mqtt_client"
-	"github.com/koupleless/virtual-kubelet/tunnel/mqtt_tunnel"
+	"github.com/koupleless/virtual-kubelet/tunnel/koupleless_mqtt_tunnel"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -21,7 +21,7 @@ func init() {
 func TestBaseProvider_Lifecycle(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	mt := &mqtt_tunnel.MqttTunnel{}
+	mt := &koupleless_mqtt_tunnel.MqttTunnel{}
 
 	err := mt.Register(ctx, "test-client", "test", nil, nil, nil, nil, nil)
 	assert.NoError(t, err)
